added sqlite database, added function for taking connection for SQLite
