+++ StudentService/build.gradle
@@ -10,8 +10,14 @@ repositories {
 }
 
 dependencies {
+    //logging
     implementation 'org.slf4j:slf4j-api:2.0.12'
+
+    //database
     implementation 'org.postgresql:postgresql:42.7.2'
+    implementation 'org.xerial:sqlite-jdbc:3.45.1.0'
+
+    //testing
     testImplementation platform('org.junit:junit-bom:5.9.1')
     testImplementation 'org.junit.jupiter:junit-jupiter'
 }
