+++ StudentService/src/main/java/org/example/service/dmlService/impl/PostgreDML.java
@@ -1,6 +1,6 @@
 package org.example.service.dmlService.impl;
 
-import org.example.component.PGConnection;
+import org.example.component.DatabaseConnection;
 import org.example.service.dmlService.DMLService;
 
 import java.sql.Connection;
@@ -20,7 +20,7 @@ public class PostgreDML implements DMLService {
 
     @Override
     public List<Integer> getAllAge() {
-        try (Connection con = PGConnection.INSTANCE.getConnection();
+        try (Connection con = DatabaseConnection.INSTANCE.getPGConnection();
              ResultSet resultSet = getResultSet(con)) {
             List<Integer> ages = new ArrayList<>();
